split FuturePerfect.scala in order to make it more readable
